LRCI-3931 Move Jenkins GitHub URL methods into JobEntity
